requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
fer==21.1.0
numpy==1.26.4
opencv-python-headless==4.8.0.74
Pillow==10.0.0
typing_extensions==4.7.1
